refactor: unify file handling service
