+++ services/__init__.py
@@ -18,12 +18,13 @@ from .registry import get_service
 from .upload_processing import UploadAnalyticsProcessor
 from .db_analytics_helper import DatabaseAnalyticsHelper
 from .summary_reporting import SummaryReporter
+from .data_processing.file_handler import FileHandler
 
 logger = logging.getLogger(__name__)
 
 # Resolve optional services from the registry
-FileProcessor = get_service("FileProcessor")
-FILE_PROCESSOR_AVAILABLE = FileProcessor is not None
+FileHandlerService = get_service("FileHandler")
+FILE_HANDLER_AVAILABLE = FileHandlerService is not None
 
 get_analytics_service = get_service("get_analytics_service")
 create_analytics_service = get_service("create_analytics_service")
@@ -31,8 +32,8 @@ AnalyticsService = get_service("AnalyticsService")
 ANALYTICS_SERVICE_AVAILABLE = AnalyticsService is not None
 
 __all__ = [
-    "FileProcessor",
-    "FILE_PROCESSOR_AVAILABLE",
+    "FileHandler",
+    "FILE_HANDLER_AVAILABLE",
     "get_analytics_service",
     "create_analytics_service",
     "AnalyticsService",
