+++ tests/test_unicode_handler_full.py
@@ -17,8 +17,8 @@ from callback_controller import (
     fire_event,
     callback_handler,
 )
-from robust_file_processor import (
-    RobustFileProcessor,
+from services.data_processing.file_handler import (
+    FileHandler as RobustFileProcessor,
     process_file_simple,
     FileProcessingError,
 )
